LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/ignore-opens-during-unload.window-expected.txt


Harness Error (TIMEOUT), message = null

PASS document.open should bail out when ignore-opens-during-unload is greater than 0 during beforeunload event (in top-level browsing context)
PASS document.open should bail out when ignore-opens-during-unload is greater than 0 during beforeunload event (open(parent) while unloading parent and child)
PASS document.open should bail out when ignore-opens-during-unload is greater than 0 during beforeunload event (open(parent) while unloading child only)
PASS document.open should bail out when ignore-opens-during-unload is greater than 0 during pagehide event (in top-level browsing context)
PASS document.open should bail out when ignore-opens-during-unload is greater than 0 during pagehide event (open(parent) while unloading parent and child)
PASS document.open should bail out when ignore-opens-during-unload is greater than 0 during pagehide event (open(parent) while unloading child only)
PASS document.open should bail out when ignore-opens-during-unload is greater than 0 during unload event (in top-level browsing context)
PASS document.open should bail out when ignore-opens-during-unload is greater than 0 during unload event (open(parent) while unloading parent and child)
PASS document.open should bail out when ignore-opens-during-unload is greater than 0 during unload event (open(parent) while unloading child only)
TIMEOUT document.open should bail out when ignore-opens-during-unload is greater than 0 during visibilitychange event (in top-level browsing context) Test timed out
TIMEOUT document.open should bail out when ignore-opens-during-unload is greater than 0 during visibilitychange event (open(parent) while unloading parent and child) Test timed out
TIMEOUT document.open should bail out when ignore-opens-during-unload is greater than 0 during visibilitychange event (open(parent) while unloading child only) Test timed out
